backend/requirements.txt
fastapi
uvicorn
python-jose[cryptography]
python-multipart
psycopg2-binary
sqlalchemy
python-dotenv
